requirements.txt
python-dotenv
langchain-groq
langchain-huggingface
langchain-community
faiss-cpu
sentence-transformers
websockets
fastapi
uvicorn
jinja2
python-multipart
